moved definition to header, fixed duplicate symbol bug
